refactor: use C for the external scanner
